requirements.txt
requests>=2.28
entsoe-py>=0.7
pandas>=2.0
streamlit>=1.30
plotly>=5.0
